alterando fontes pt3
