+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Jarvis">
           <activity
-            android:name=".MainActivityDrawer" >
+            android:name=".MainActivity" >
 <!--            android:theme="@style/Theme.MaterialComponents.NoActionBar"-->
 
             <intent-filter>
