--- app/src/main/java/com/e/jarvis/FavoritosComicsFragment.kt
@@ -1,24 +1,0 @@
-package com.e.jarvis
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-class FavoritosComicsFragment : Fragment() {
-
-
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favoritos_comics, container, false)
-    }
-
-    companion object {}
-}
