+++ app/src/main/java/com/e/jarvis/MainActivity.kt
@@ -20,7 +20,7 @@ import androidx.navigation.Navigation
 import androidx.navigation.fragment.findNavController
 import kotlinx.android.synthetic.main.fragment_home.*
 
-class MainActivityDrawer : AppCompatActivity() {
+class MainActivity : AppCompatActivity() {
 
     private lateinit var appBarConfiguration: AppBarConfiguration
