+++ app/src/main/java/com/e/jarvis/ui/cadastro/CadastroFragment.kt
@@ -1,10 +1,11 @@
-package com.e.jarvis
+package com.e.jarvis.ui.cadastro
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.e.jarvis.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [FavoritosStoriesFragment.newInstance] factory method to
+ * Use the [CadastroFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class FavoritosStoriesFragment : Fragment() {
+class CadastroFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,7 @@ class FavoritosStoriesFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favoritos_stories, container, false)
+        return inflater.inflate(R.layout.fragment_cadastro, container, false)
     }
 
     companion object {
@@ -44,12 +45,12 @@ class FavoritosStoriesFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment FavoritosStoriesFragment.
+         * @return A new instance of fragment CadastroFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            FavoritosStoriesFragment().apply {
+            CadastroFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
