+++ app/src/main/java/com/e/jarvis/ui/exibe/chars/ExibeCharFragment.kt
@@ -1,4 +1,4 @@
-package com.e.jarvis
+package com.e.jarvis.ui.exibe.chars
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,9 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
-import kotlinx.android.synthetic.main.fragment_exibe_personagem.view.*
+import com.e.jarvis.R
+import kotlinx.android.synthetic.main.item_exibe.view.*
 
-class ExibePersonagemFragment : Fragment() {
+
+class ExibeCharFragment : Fragment() {
 
 
     override fun onCreateView(
@@ -16,24 +18,20 @@ class ExibePersonagemFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_exibe_personagem, container, false)
+        val view = inflater.inflate(R.layout.fragment_exibe_char, container, false)
 
-        view.btn_frag_personagem_series.setOnClickListener {
+        view.btn_exibe_series.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_personagem_to_exibe_series_fragment)
         }
 
-        view.btn_frag_personagem_comics.setOnClickListener {
+        view.btn_exibe_comics.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_personagem_to_exibe_comics_fragment)
         }
 
-        view.btn_frag_personagem_stories.setOnClickListener {
+        view.btn_exibe_stories.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_personagem_to_exibe_stories_fragment)
         }
-
-
         return view
-
-
     }
