+++ app/src/main/java/com/e/jarvis/ui/exibe/serie/ExibeSeriesFragment.kt
@@ -1,4 +1,4 @@
-package com.e.jarvis
+package com.e.jarvis.ui.exibe.serie
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,8 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.Navigation
-import kotlinx.android.synthetic.main.fragment_exibe_personagem.view.*
-import kotlinx.android.synthetic.main.fragment_exibe_series_fragment.view.*
+import com.e.jarvis.R
+import kotlinx.android.synthetic.main.item_exibe.view.*
 
 class ExibeSeriesFragement : Fragment() {
 
@@ -16,22 +16,19 @@ class ExibeSeriesFragement : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_exibe_series_fragment, container, false)
+        val view = inflater.inflate(R.layout.fragment_exibe_series, container, false)
 
-        view.btn_frag_series_char.setOnClickListener {
+        view.btn_exibe_char.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_series_to_personagem_fragment)
         }
 
-        view.btn_frag_series_comics.setOnClickListener {
+        view.btn_exibe_comics.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_series_to_comics_fragment)
         }
 
-        view.btn_frag_series_stories.setOnClickListener {
+        view.btn_exibe_stories.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_series_to_stories_fragment)
         }
-
-
-
 
         return view
     }
