+++ app/src/main/java/com/e/jarvis/ui/exibe/story/ExibeStoriesFragment.kt
@@ -1,4 +1,4 @@
-package com.e.jarvis
+package com.e.jarvis.ui.exibe.story
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,8 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.Navigation
-import kotlinx.android.synthetic.main.fragment_exibe_comics.view.*
-import kotlinx.android.synthetic.main.fragment_exibe_stories.view.*
+import com.e.jarvis.R
+import kotlinx.android.synthetic.main.item_exibe.view.*
 
 class ExibeStoriesFragment : Fragment() {
 
@@ -18,19 +18,17 @@ class ExibeStoriesFragment : Fragment() {
         // Inflate the layout for this fragment
         val view = inflater.inflate(R.layout.fragment_exibe_stories, container, false)
 
-        view.btn_frag_stories_char.setOnClickListener {
+        view.btn_exibe_char.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_stories_to_personagem_fragment)
         }
 
-        view.btn_frag_stories_series.setOnClickListener {
+        view.btn_exibe_series.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_stories_to_series_fragment)
         }
 
-        view.btn_frag_stories_comics.setOnClickListener {
+        view.btn_exibe_comics.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.navigate_stories_to_comics_fragment)
         }
-
-
 
         return view
     }
