+++ app/src/main/java/com/e/jarvis/ui/favorites/FavoritosComicsFragment.kt
@@ -1,13 +1,14 @@
-package com.e.jarvis
+package com.e.jarvis.ui.favorites
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.e.jarvis.R
 
 
-class FavoritosPersonagemFragment : Fragment() {
+class FavoritosComicsFragment : Fragment() {
 
 
 
@@ -17,7 +18,7 @@ class FavoritosPersonagemFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favoritos_personagem, container, false)
+        return inflater.inflate(R.layout.fragment_favoritos_comics, container, false)
     }
 
     companion object {}
