+++ app/src/main/java/com/e/jarvis/ui/favorites/FavoritosSeriesFragment.kt
@@ -1,10 +1,11 @@
-package com.e.jarvis
+package com.e.jarvis.ui.favorites
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.e.jarvis.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
