+++ app/src/main/java/com/e/jarvis/ui/login/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.e.jarvis
+package com.e.jarvis.ui.login
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,8 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.Navigation
+import com.e.jarvis.R
 import kotlinx.android.synthetic.main.fragment_login.view.*
-import kotlinx.android.synthetic.main.fragment_quiz.view.*
 
 class LoginFragment : Fragment() {
 
@@ -22,7 +22,7 @@ class LoginFragment : Fragment() {
     ): View? {
         // Inflate the layout for this fragment
         val view = inflater.inflate(R.layout.fragment_login, container, false)
-        view.bnewuser.setOnClickListener {
+        view.btn_login_newUser.setOnClickListener {
             Navigation.findNavController(view).navigate(R.id.action_loginFragment_to_cadastroFragment)
         }
         return view
