+++ app/src/main/java/com/e/jarvis/ui/search/PesquisaFragment.kt
@@ -1,4 +1,4 @@
-package com.e.jarvis
+package com.e.jarvis.ui.search
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
+import com.e.jarvis.R
 import kotlinx.android.synthetic.main.fragment_pesquisa.view.*
 
 
@@ -18,10 +19,9 @@ class PesquisaFragment : Fragment() {
         // Inflate the layout for this fragment
         val view = inflater.inflate(R.layout.fragment_pesquisa, container, false)
 
-        view.barra_pesquisa2.setOnClickListener{
+        view.sv_search.setOnClickListener{
             Navigation.findNavController(view).navigate(R.id.navigate_to_exibe_personagem_fragment)
         }
-
         return view
     }
