+++ app/src/main/res/layout/fragment_cadastro.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+<FrameLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bkg2">
@@ -10,15 +10,12 @@
         android:layout_height="match_parent">
 
         <LinearLayout
-            xmlns:android="http://schemas.android.com/apk/res/android"
-            xmlns:app="http://schemas.android.com/apk/res-auto"
-            xmlns:tools="http://schemas.android.com/tools"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:orientation="vertical"
             >
             <EditText
-                android:id="@+id/editTextTextEmailAddressCadastro"
+                android:id="@+id/et_cadastro_email"
                 style="@style/fonteSecundaria"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -32,7 +29,7 @@
                 android:textColorHint="@color/white" />
 
             <EditText
-                android:id="@+id/editTextTextPasswordCadastro"
+                android:id="@+id/et_cadastro_password"
                 style="@style/fonteSecundaria"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -45,7 +42,7 @@
                 android:textColorHint="@color/white" />
 
             <EditText
-                android:id="@+id/editTextTextConfirmPasswordCadastro"
+                android:id="@+id/et_cadastro_password_confirm"
                 style="@style/fonteSecundaria"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -58,7 +55,7 @@
                 android:textColorHint="@color/white" />
 
             <EditText
-                android:id="@+id/editTextDate"
+                android:id="@+id/et_cadastro_birth"
                 style="@style/fonteSecundaria"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -71,7 +68,7 @@
                 android:textColorHint="@color/white" />
 
             <Button
-                android:id="@+id/register"
+                android:id="@+id/btn_cadastro_register"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="11dp"
