+++ app/src/main/res/layout/fragment_exibe_char.xml
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bkg1"
     android:orientation="vertical">
+
 
     <include layout="@layout/item_exibe" />
 
@@ -17,28 +19,29 @@
         app:cardCornerRadius="20dp"
         app:strokeColor="@color/black"
         app:strokeWidth="3dp">
+
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical">
             <TextView
-                android:id="@+id/tv_titulo_frag_comics"
-                style="@style/fontePrincipal"
+                android:id="@+id/tv_titulo_frag_char"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:layout_margin="15dp"
                 android:gravity="center"
-                android:text="Nome da Comic"
+                android:text="Nome do Personagem"
                 android:textColor="@color/titleColor"
+                style="@style/fontePrincipal"
                 android:textSize="24sp" />
             <TextView
-                android:id="@+id/tv_descricao_frag_comics"
-                style="@style/fonteLeitura"
+                android:id="@+id/tv_descricao_frag_personagem"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_margin="20dp"
                 android:text="texto descrição"
                 android:textColor="@color/black"
+                style="@style/fonteLeitura"
                 android:textSize="16sp" />
         </LinearLayout>
     </com.google.android.material.card.MaterialCardView>
