+++ app/src/main/res/layout/fragment_favoritos_series.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@drawable/q1_bw"
-    tools:context=".FavoritosSeriesFragment">
+    tools:context=".ui.favorites.FavoritosSeriesFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
