+++ app/src/main/res/layout/fragment_favoritos_stories.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@drawable/q1_bw"
-    tools:context=".FavoritosStoriesFragment">
+    tools:context=".ui.favorites.FavoritosStoriesFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
