+++ app/src/main/res/layout/fragment_home.xml
@@ -5,125 +5,48 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bkg4"
-    android:gravity="center"
+    android:orientation="vertical">
 
-    >
-
-    <androidx.constraintlayout.widget.ConstraintLayout
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center">
-
-        <SearchView
-            android:id="@+id/barra_pesquisa1"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="40dp"
-            android:background="@drawable/shape_search"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.5"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <!--        <androidx.recyclerview.widget.RecyclerView-->
-        <!--            android:layout_width="match_parent"-->
-        <!--            android:layout_height="450dp"-->
-        <!--            android:gravity="center"-->
-        <!--            app:layout_constraintBottom_toBottomOf="parent"-->
-        <!--            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"-->
-        <!--            tools:listitem="@layout/item_top_heroi"/>-->
-
-
-        <!-- visualização de como ficaria o top 9 -->
-
-        <TextView
-            android:id="@+id/home_mais_procurados"
-            style="@style/fontePrincipal"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="36dp"
-            android:text="@string/top_personagens"
-            android:textColor="@color/white"
-            android:textSize="20sp"
-            android:textStyle="bold"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.496"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/barra_pesquisa1" />
+        android:layout_height="match_parent">
 
         <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="450dp"
-            android:layout_marginTop="24dp"
-            android:gravity="center"
-            android:orientation="vertical"
-            android:paddingRight="15dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/home_mais_procurados">
+            android:layout_height="match_parent"
+            android:orientation="vertical">
 
-            <LinearLayout
+            <SearchView
+                android:id="@+id/sv_home"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="50dp"
+                android:layout_marginTop="50dp"
+                android:layout_marginEnd="50dp"
+                android:background="@drawable/shape_search" />
+
+            <TextView
+                android:id="@+id/tv_home_top"
+                style="@style/fontePrincipal"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:orientation="horizontal">
+                android:layout_gravity="center"
+                android:layout_marginTop="20dp"
+                android:text="@string/top_personagens"
+                android:textColor="@color/white"
+                android:textSize="20sp"
+                android:textStyle="bold" />
 
-                <include
-                    android:id="@+id/top1"
-                    layout="@layout/item_top_heroi" />
-
-                <include
-                    android:id="@+id/top2"
-                    layout="@layout/item_top_heroi" />
-
-                <include
-                    android:id="@+id/top3"
-                    layout="@layout/item_top_heroi" />
-
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="wrap_content"
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/rc_home"
+                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:orientation="horizontal">
-
-                <include
-                    android:id="@+id/top4"
-                    layout="@layout/item_top_heroi" />
-
-                <include
-                    android:id="@+id/top5"
-                    layout="@layout/item_top_heroi" />
-
-                <include
-                    android:id="@+id/top6"
-                    layout="@layout/item_top_heroi" />
-
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal">
-
-                <include
-                    android:id="@+id/top7"
-                    layout="@layout/item_top_heroi" />
-
-                <include
-                    android:id="@+id/top8"
-                    layout="@layout/item_top_heroi" />
-
-                <include
-                    android:id="@+id/top9"
-                    layout="@layout/item_top_heroi" />
-
-            </LinearLayout>
-
+                android:layout_marginTop="20dp"
+                android:gravity="center"
+                app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+                app:spanCount="3"
+                tools:itemCount="9"
+                tools:listitem="@layout/item_top_heroi" />
         </LinearLayout>
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-
+    </ScrollView>
 </LinearLayout>
