+++ app/src/main/res/layout/fragment_pesquisa.xml
@@ -1,54 +1,39 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".PesquisaFragment"
     android:background="@drawable/bkg6"
-    android:windowSoftInputMode="stateAlwaysVisible" >
+    android:orientation="vertical"
+    >
 
-    <androidx.constraintlayout.widget.ConstraintLayout
+    <SearchView
+        android:background="@drawable/shape_search"
+        android:id="@+id/sv_search"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="50dp"
+        android:layout_marginEnd="50dp"
+        android:layout_marginStart="50dp"
+        android:layout_width="match_parent" />
+
+    <TextView
+        style="@style/fonteSecundaria"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
         android:layout_width="match_parent"
-        android:layout_height="400dp"
-        >
+        android:gravity="center"
+        android:text="@string/personagens_encontrados"
+        android:textColor="@color/white"
+        android:textSize="18sp" />
 
-        <SearchView
-            android:id="@+id/barra_pesquisa2"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="40dp"
-            android:background="@drawable/shape_search"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.5"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="108dp"
-            android:text="@string/personagens_encontrados"
-            android:textColor="@color/white"
-            android:textSize="18sp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rcv_pesquisa"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.8"
-            tools:layout_editor_absoluteX="0dp"
-            tools:listitem="@layout/item_search_heroi" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-
-
-</FrameLayout>
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rv_search"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:layout_width="match_parent"
+        android:orientation="horizontal"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+        tools:listitem="@layout/item_search_heroi" />
+</LinearLayout>
