+++ app/src/main/res/layout/fragment_question.xml
@@ -17,6 +17,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
         <ImageView
+            android:id="@+id/iv_question"
             android:layout_width="match_parent"
             android:layout_height="200dp"
             android:src="@drawable/ic_question"
@@ -37,7 +38,7 @@
                 android:orientation="vertical"
                 android:padding="15dp">
                 <TextView
-                    android:id="@+id/txtQuestion"
+                    android:id="@+id/tv_question"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:justificationMode="inter_word"
@@ -55,28 +56,28 @@
                     android:orientation="horizontal"
                     android:padding="20dp">
                     <TextView
-                        android:id="@+id/txtAlt1"
+                        android:id="@+id/tv_question_alt1"
                         android:layout_width="150dp"
                         android:layout_height="wrap_content"
                         android:textColor="@color/black"
                         style="@style/fonteLeitura"
                         android:text="A - Teste" />
                     <TextView
-                        android:id="@+id/txtAlt2"
+                        android:id="@+id/tv_question_alt2"
                         android:layout_width="150dp"
                         android:layout_height="wrap_content"
                         android:textColor="@color/black"
                         style="@style/fonteLeitura"
                         android:text="B - Teste" />
                     <TextView
-                        android:id="@+id/txtAlt3"
+                        android:id="@+id/tv_question_alt3"
                         android:layout_width="150dp"
                         android:layout_height="wrap_content"
                         android:textColor="@color/black"
                         style="@style/fonteLeitura"
                         android:text="C - Teste" />
                     <TextView
-                        android:id="@+id/txtAlt4"
+                        android:id="@+id/tv_question_alt4"
                         android:layout_width="150dp"
                         android:layout_height="wrap_content"
                         android:textColor="@color/black"
