+++ app/src/main/res/layout/fragment_quiz.xml
@@ -34,7 +34,7 @@
                 android:padding="15dp">
 
                 <TextView
-                    android:id="@+id/txtQuizTitle"
+                    android:id="@+id/tv_quiz_title"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:text="@string/QuizTitle"
@@ -46,7 +46,7 @@
                     />
 
                 <TextView
-                    android:id="@+id/txtQuizIntro"
+                    android:id="@+id/tv_quiz_intro"
                     android:justificationMode="inter_word"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -73,7 +73,7 @@
                 android:padding="15dp">
 
                 <TextView
-                    android:id="@+id/txtPTitle"
+                    android:id="@+id/tv_quiz_progress"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     style="@style/fontePrincipal"
@@ -90,7 +90,7 @@
                     android:orientation="horizontal">
 
                     <ImageView
-                        android:id="@+id/ivPIcon"
+                        android:id="@+id/iv_quiz_progress"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_marginLeft="30dp"
@@ -101,7 +101,7 @@
                         />
 
                     <TextView
-                        android:id="@+id/PRank"
+                        android:id="@+id/iv_quiz_rank"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         style="@style/fonteLeitura"
@@ -132,7 +132,7 @@
                         android:progressTint="#33ff99" />
 
                     <TextView
-                        android:id="@+id/txtProgress"
+                        android:id="@+id/tv_quiz_progressPerc"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:textColor="@color/black"
@@ -144,7 +144,7 @@
                     android:layout_height="wrap_content"
                     android:orientation="vertical">
                     <TextView
-                        android:id="@+id/txtProgressAch"
+                        android:id="@+id/tv_quiz_history"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:textColor="@color/black"
@@ -157,7 +157,7 @@
         </com.google.android.material.card.MaterialCardView>
 
         <Button
-            android:id="@+id/btnQuiz"
+            android:id="@+id/btn_quiz"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="20dp"
@@ -167,8 +167,5 @@
             android:text="@string/StartQuiz"
             style="@style/fontePrincipal"
             />
-
     </LinearLayout>
-
-
 </FrameLayout>
