+++ app/src/main/res/layout/item_exibe.xml
@@ -1,0 +1,75 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
+    >
+
+    <ImageView
+        android:id="@+id/iv_exibe_menu"
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        android:scaleType="fitXY"
+        android:src="@drawable/wolverine"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <LinearLayout
+        android:id="@+id/ll_frag_comics"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/black"
+        android:orientation="horizontal"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/iv_exibe_menu">
+
+        <Button
+            android:id="@+id/btn_exibe_char"
+            style="@style/fontePrincipal"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:backgroundTint="@color/black"
+            android:drawableTop="@drawable/ic_btn_personagem"
+            android:text="Char."
+            android:textColor="@color/white" />
+
+        <Button
+            android:id="@+id/btn_exibe_series"
+            style="@style/fontePrincipal"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:backgroundTint="@color/black"
+            android:drawableTop="@drawable/ic_btn_series"
+            android:text="Series"
+            android:textColor="@color/white" />
+
+        <Button
+            android:id="@+id/btn_exibe_comics"
+            style="@style/fontePrincipal"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:backgroundTint="@color/black"
+            android:drawableTop="@drawable/ic_btn_comics"
+            android:text="Comics"
+            android:textColor="@color/white" />
+
+        <Button
+            android:id="@+id/btn_exibe_stories"
+            style="@style/fontePrincipal"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:backgroundTint="@color/black"
+            android:drawableTop="@drawable/ic_btn_stories"
+            android:text="Stories"
+            android:textColor="@color/white" />
+
+    </LinearLayout>
+</LinearLayout>
