--- app/src/main/res/navigation/exibe_navigation.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/exibe_navigation"
-    app:startDestination="@id/exibePersonagemFragment">
-
-</navigation>
