+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -17,9 +17,9 @@
 
     <fragment
         android:id="@+id/exibePersonagemFragment"
-        android:name="com.e.jarvis.ExibePersonagemFragment"
+        android:name="com.e.jarvis.ui.exibe.chars.ExibeCharFragment"
         android:label="NOME DO PERSONAGEM"
-        tools:layout="@layout/fragment_exibe_personagem">
+        tools:layout="@layout/fragment_exibe_char">
         <action
             android:id="@+id/navigate_personagem_to_exibe_series_fragment"
             app:destination="@id/exibeSeriesFragement" />
@@ -32,7 +32,7 @@
     </fragment>
     <fragment
         android:id="@+id/exibeComicsFragment"
-        android:name="com.e.jarvis.ExibeComicsFragment"
+        android:name="com.e.jarvis.ui.exibe.comic.ExibeComicsFragment"
         android:label="NOME DO PERSONAGEM"
         tools:layout="@layout/fragment_exibe_comics">
         <action
@@ -47,7 +47,7 @@
     </fragment>
     <fragment
         android:id="@+id/exibeStoriesFragment"
-        android:name="com.e.jarvis.ExibeStoriesFragment"
+        android:name="com.e.jarvis.ui.exibe.story.ExibeStoriesFragment"
         android:label="NOME DO PERSONAGEM"
         tools:layout="@layout/fragment_exibe_stories">
         <action
@@ -62,7 +62,7 @@
     </fragment>
     <fragment
         android:id="@+id/pesquisaFragment"
-        android:name="com.e.jarvis.PesquisaFragment"
+        android:name="com.e.jarvis.ui.search.PesquisaFragment"
         android:label="JARVIS DB"
         tools:layout="@layout/fragment_pesquisa">
         <action
@@ -71,7 +71,7 @@
     </fragment>
     <fragment
         android:id="@+id/exibeSeriesFragement"
-        android:name="com.e.jarvis.ExibeSeriesFragement"
+        android:name="com.e.jarvis.ui.exibe.serie.ExibeSeriesFragement"
         android:label="NOME DO PERSONAGEM">
         <action
             android:id="@+id/navigate_series_to_personagem_fragment"
@@ -85,7 +85,7 @@
     </fragment>
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.e.jarvis.LoginFragment"
+        android:name="com.e.jarvis.ui.login.LoginFragment"
         android:label="Login"
         tools:layout="@layout/fragment_login">
         <action
@@ -103,12 +103,12 @@
     </fragment>
     <fragment
         android:id="@+id/favoritosPersonagemFragment"
-        android:name="com.e.jarvis.FavoritosPersonagemFragment"
+        android:name="com.e.jarvis.ui.favorites.FavoritosPersonagemFragment"
         android:label="Favoritos"
         tools:layout="@layout/fragment_favoritos_personagem" />
     <fragment
         android:id="@+id/cadastroFragment"
-        android:name="com.e.jarvis.CadastroFragment"
+        android:name="com.e.jarvis.ui.cadastro.CadastroFragment"
         android:label="New user"
         tools:layout="@layout/fragment_cadastro" />
     <fragment
